Send emails after the creation of a sharing
This PR introduces the following behavior: when the user creates a new
sharing, "invitation emails" are automatically sent to the recipients.

To send the emails a job is queued for each recipient.

This email contains a button that, once clicked, will trigger the next
step of the sharing process.

A route (PUT /sharings/id/sendMails) was also introduced in case the
email could not be sent to one or more of the recipients. In addititon
the status of the recipients for whom the email was not sent is changed
to "error".

To allow all this, it was necessary to modify the structure of the
former `sharingRecipient` so that one could access the underlying
`Recipient`. As a consequence, and to add clarity, it was renamed
`RecipientStatus`.