One OAuth client per sharing
Before this PR the behavior regarding OAuth in C2C sharing was to
register a sharer once at each recipient and then ask for an access
token for every new share using the same OAuth client. We also thought
that a recipient could revoke an access token directly, which is not
true. A recipient can only revoke an OAuth client, which means all the
access tokens and this is not what we want.

To give the recipient the possibility to revoke a sharing we have to
have one OAuth client per sharing. This is what this PR introduces.

The modifications can be roughly summed up as following:
* Refine the sharing status to include errors when the sharer could not
register or when the email was not sent.
* Move the `RecipientStatus` code in the "Recipients.go" file.
* Add `auth.Client` and `auth.AccessToken` structures in
`RecipientStatus` so as to access the OAuth information later on.
* Creating a recipient no longer triggers the registration process.
* Creating a sharing now triggers the registration process for each
recipient. Their status is updated accordingly.
* Fix lots of tests that these modifications broke.
* Fix gometalinter numerous warnings…