Use same `createdAt` and `updatedAt` dates
Before this PR the recipients and the sharer would have different dates
for the shared files: the recipients would set the `createdAt` date to
when they received a file, and `updatedAt` to when they received an
update.

Now both the sharer and the recipients have the same `createdAt` and
`updatedAt` dates.