Prevent infinite loop after update on shared document
In a master-master sharing, all updates are propagated from and to the
sharer and recipients. So, an update from the sharer would be propagated
to the recipient, seen as a new update on the recipient side, propagated
once again
to the sharer, etc.

To prevent this, we compare the local and remote md5 of each document.
If nothing had changed excpect the revision, we do not propagate the
update.