feat: folders hierarchy in sharing
This PR introduces the necessary code to propagate what we call the
"folders hierarchy": if Alice shares a folder containing subfolders then
the recipient should retrieve the same structure, and if changes are
made they should also be propagated.

In addition this PR introduces the possibility to create specific
folders other than "Shared With Me". It will be used for photos in
particular.