feature: trash document when sharing ends
With this PR when a document is no longer shared it is trashed. If the
document is still shared in another sharing it is kept.
To enable this feature the `Valid` method of the JSONDoc objet was
improved to handle `referenced_by` permissions.

Several other additions where also made:
* some logging was added;
* the test coverage was improved in pkg/workers/sharings;
* if we share a JSON document (not an update) a check is performed on
its doctype to see if the database already exists and, if it doesn't, to
create it;
* in web/files/referencedby.go the function `RemoveReferencedHandler`
returns `c.JSON` instead of `c.NoContent` because the latter triggers
the error "http: request method or response status code does not allow
body".