feat: a document can be reshared
With this PR it is possible to share several times the same document. We
detect that a document is reshared when we try to fetch it: we receive a
403 error — the document already exists but we can't access it.
When we receive such error we send the document as is if it was a
creation and the recipient will interpret it as an update.