fix: prevent consecutive sendings
When Alice shares a photo with Bob through a Master-Master sharing,
Bob's Cozy starts the sharing_update worker. It sees this event as a
creation and tries to send that very same photo to Alice. This led to
errors in the stack log.

This PR fixes this behavior: when we send a file we first try to get it
and unless we receive a 404 or 403 error we abort.