Update konnectors workflow with aggregator informations (#1667)
Following https://github.com/cozy/cozy-stack/pull/1665

![konnector-workflow](https://user-images.githubusercontent.com/776764/50104379-25594480-022a-11e9-9149-e3d141a0a8bc.png)

[...]

#### Aggregator accounts

Some konnectors are based on a aggregator service. An aggregator is declared
in the konnector manifest and specify an `accountId` property. When Cozy-Home
detects this property, it check if an account with this given id exists. If not,
it creates it. Every account created for a konnector based on an aggregator is
then related to this aggregator account. This relationship is called `master`
account.

```json
{
    "auth": {
        "login": "000000000",
        "password": "**********"
    },
    "folderPath": "/Administratif/Bankbank",
    "label": "Bankbank",
    "namePath": "Bankbank",
    "relationships": {
      "master": {
        "_id": "service-aggregator-account",
        "_type": "io.cozy.accounts"
      }
    }
}
```
[...]