Fix for same app/version but not code from 2 registries (#1700)
When an app is available on two registry spaces, at the same version
number, but not with the same code (customized app for example), the
stack was serving the code from the first version installed. With
this fix, we can differentiate these cases and serve the right code
for the webapp/konnector.