Deprecate the dev flag on instances
A stack can be built in two release modes: dev and prod. We also
has a flag per instance, but it was confusing to have both and, in
practice, using a prod release with a dev instance didn't work.

Note that a prod release can be used as a dev release by starting the
server like this:

	$ cozy-stack serve --dev