Refactor to broke the instance package in two
Before this change, the jobs package was imported by the instance
package. But I want to access the instance from the jobs package. So,
I've split the instance package in two to broke the circular import
issue. In ./pkg/instance, we have kept the Instance struct with the
common functions to use it. And, we moved all the methods that change
the instance (creation, patch, deletion, changing passphrase or 2FA) in
the new ./pkg/instance/lifecycle package.