Split the apps package in two
We want to make the jobs package able to import the instance package.
But the instance package depends of the apps package that depends of the
jobs package. It would be a circular import issue. To fix it, we have
extracted the copier and file server parts of the apps in a new package:
./pkg/apps/appfs.