Use a map for contacts
We were using structs for the contacts doctype, but it forces to have
exactly the same mapping as the front applications. It can be dangerous
and lead to data loss. For example, if a contact was added to a group,
and she accepts a sharing, the contact will be removed from the group.