Add the fallback URI for authorize with linked apps
When a mobile app is linked to a web app, the stack will add a fallback
parameter on the redirection after the authorize step of OAuth. This
fallback parameter is the URL of the web app on this instance. It will
allow to send the user to this URL if the mobile app can't be opened.