Block disk-quota requests with no token
Before this change, the stack was checking only if the user was logged
in. But, an attacker that knows the URL of the cozy can make a cross
site request forgery on this endpoint. Now, the stack also checks that a
web app token is also present. It should be compatible with the
existant, as the cozy-bar was already sending this token.