Allow mobile apps to install app
For cozy instances in a context where allow_install_via_a_permission is
activated, the stack should allow a mobile app with the right
permissions to install other apps. To do that, we need to check OAuth
permissions documents on the endpoint for listing apps in the registry
and installing an application.