Fix cozy-app-dev container start-up behind proxy
When behind a corporate proxy, the start-up script inside the Docker
container used curl to query ":8080", which cannot match any no_proxy
item, and thus systematically goes through the corporate proxy when one
is configured, resulting in a start-up failure. This commit fixes this
query so that it succeeds in that kind of environment as long as
"localhost" is defined in no_proxy.