Make cozy-stack instances ls faster
Listing instances is done by fetching the instance information from
CouchDB with pagination. By fetching 100 instances per page, we are
making a lot of calls, and each call has a fixed latency that makes the
whole slow. We want to keep to pagination to avoid timeouts, but we can
increase the number of items per page to 10000 to make the listing
faster.