Remove the go-git dependency
At the beginning, we were using the go-git library to install
applications from the git repository. It was not ideal as this library
was not capable to handle the tricky cases like the git binary (using
something else that master for the default branch for example). So, we
moved to use the git binary by default, and with a fallback to go-git
when the binary was not available.

Now, most applications are installed from the registry and don't need
git at all, and for the developers, we can expect that they have a git
binary executable from the stack. So, there is no need to keep to go-git
library, and one less library is good thing: less weight, less
maintenance.