Allow to trigger a job on a changed value
Add a special `!=` for trigger values to indicate that a job should be
triggered only the value for the selected field has changed on the
document.

For example, a trigger on `io.cozy.bank.operations:UPDATED:!=:category`
will be fired only if the category has changed.

See https://github.com/cozy/cozy-banks/pull/1508