Fix renaming a file from NFC to NFD
When a file is renamed, its name is different, but it can be very
subtle. For example, it can have the same unicode meaning but just
with two different normalizations, like NFC and NFD. CouchDB does
a unicode normalization when using strings in mango selector, so
we had to make the selector more robust for this case to avoid a
loop where the same sub-directories where renamed.