Improve password reset flow on mobile
On mobile, if the user asks for a password reset, the redirect param was
lost, and when the user logs in, they was redirected to the home app
instead of being sent to the authorize page. The redirect param is now
transfered from the login page to the reset page, and then from the
reset page to the login page.