Fix a special case of moving/trashing a directory (#2104)
When a directory is moved or trashed, if another directory has the same
name but not the same case (like /Photos and /PHOTOS), the VFS could be
corrupted. The mango selector `$lt` and `$gt` of CouchDB can give
suprising results when filtering documents[1], and the query to find the
children of a folder can give bad results because of that. This commit
add a check in go to fix this issue.

[1]: http://docs.couchdb.org/en/stable/ddocs/views/collation.html#collation-specification