Delete versions asynchronously when emptying the trash
With swift layout v3, the files are versioned. When the trash is
emptied, the files are deleted from CouchDB during the HTTP request, and
a job is pushed for trash-files to clean the rest later. This worker
will clean the file versions in CouchDB, and the files and their
versions in Swift. It should help to make the request to empty the trash
a lot faster.