Fix accounts cleaning on instance deletion
When an instance is destroyed, the stack ensures that io.cozy.accounts
for a konnector with on_delete_account are deleted, and the associated
job has finished in success. It allows to clean the accounts on remote
systems. In particular, this commit takes care to not delete aggregator
accounts before the other accounts have been cleaned as they can be
necessary to authenticate on the remote systems.