Fix versioning with shared files
When a file is shared, and is modified with both its content and its
name changed, the sharing makes two updates in CouchDB. When the
destination instance is in Swift layout v3, we need to take care that
the internalID used for the new version on the first change is also
used for the second change. It was not the case, and fixing it avoids
FSCK mismatches between the versions in CouchDB and Swift.