Fix loops in sharing with Swift layout v3
If a conflict happens with a shared file, the ID of the file was set to
a new value with Swift layout v3. As such, it was seen as a new file,
and the sharing will upload it on the other side. But the other side has
already the original file at this location, and it created a new
conflict, and the same issue happens again. This starts creating
conflicts in a loop, which is very bad. This commit fixes this.