Fix sending push notifications
When there are many OAuth clients registered on an instance, but only a
few of them has a notification_device_token, the stack can miss sending
some push notifications. This issue comes from the pagination on the
_find request to find the notifiable clients. By including a mango
selector to find only the OAuth clients with a
notification_device_token, we can avoid this issue.