Improve the lock with redis
The package pkg/lock has two implementations, one with an in-process
mutex, and the other with redis. The redis lock was not complete, and it
didn't check anything when calling RLock(). It has been fixed.

Note that the current code does not try to allow two cozy-stacks to
share a lock for reading. If one cozy-stack has locked for reading a
lock, another cozy-stack will have to wait that the lock has been
released before being able to give a lock for reading on the same name.
It may be improved, but I prefer to err on the safe side for now.