Use Max-Age for the cookie in the CSRF middleware (#2208)
Using Expires for the cookie in the CSRF middleware can lead to some
issues, that Max-Age doesn't have. With Expires, the server defines the
expiration date, and this date will be interpreted by the client. The
client and the server can be in different timezones, or the client can
be misconfigured (no NTP). If the MaxAge configuration from the
middleware is short (let's say hours, not days), it can happen that some
clients throw the cookie as expired when they received it.

To do that, we have copied the CSRF middleware from echo, and made a
pull request for this change upstream. See
https://github.com/labstack/echo/pull/1450