Improve rate limiting with redis (#2250)
The rate limiting works with counters. The redis implementation has a
TTL set on the counters to reset the counter after a period has elapsed.
But, it worked with 2 calls to redis: the first to increment the
counter, and the second to set the TTL if the counter has been created.
Network errors may happen between the two calls, and we can have
counters with no TTL. In that case, the counter will grow but will never
be reset, which is not good. This commit fixes this issue by using a
single call to redis with a lua script.