Remove triggers from scheduling when rate limited (#2271)
When the stack is using redis for jobs and triggers, it needs to remove
the triggers from the scheduling key when a rate limit is reached. It
avoids that to trying several times to push a job for the rate limited
trigger, and should help to keep a healthy state of the global
environment.