Accept . at the end of short sharecode
In theory, a short shortcode is exactly 12 characters. But somethimes,
when people shares a public link with this token, they can put a "."
just after the link to finish their sentence, and this "." can be added
to the token. So, it's better to accept a shortcode with a final ".",
and clean it.