Improve how the folder for konnectors is recreated
When a konnector needs a folder to put files, this folder can be created
by harvest with a specific name, and its ID is saved in the trigger. If
this folder is moved, the ID allows to find it again. But if the folder
is deleted, the stack needs to recreate the folder, and update the
trigger with the new ID. There is also some other things, liking putting
a reference on the folder to the konnector, and executing the konnector
with the updated folder. The name of this folder is taken from the
defaultFolderName of the account, but if it is empty, it can be the
folderName attribute (for legacy reasons) or a name can be computed from
the konnector slug.