Add a deleting flag on instance
When deleting an instance, the first step is to clean the accounts. Tis
step must not run twice in parallel if the instance deletion is runned
twice. And if it fails, it must be manually managed. So, we have added a
`deleting` flag on the instance to avoid those issues.