Tweak the Cozy to Cozy sharing parameters
When we have seen in the logs some cases where there a lot of failures
of the sharing workers in a short period of time, related to the lock.
They probably come of an accumulation of retry jobs. This commit tries
to tweak the parameters to avoid this cascading failure:

- decrease the number of retries
- increase the initial backoff period
- increase the debounce for the share-upload trigger
- increase the sequence number if the last try for uploading a file
  fails (to allow to skip a corrupted file).