Fix sharing a single note from Drive web (#2361)
When the Drive web app shares a single note to a set of recipients in
read-write mode, it creates a sharing with a rule:
- add: none
- update: sync
- delete: revoke.
With this rule, the stack didn't create an io.cozy.shared for the note
on the recipient instance, and thus, the note wasn't seen as part of the
sharing. It means that later updates of this note made on the sharer
instance are not accepted by the recipients, and the open route for
notes sent parameters as if the note was local, not shared. With this
commit, the io.cozy.shared is correctly created, which fixes our issues.