Fix an issue when a file is uploaded twice (#2363)
In VFS Afero and Swift layout v3, if the same file is uploaded twice in
parallel, it may corrupt the instance (with a duplicate file). The check
to ensure the unicity of the path is made at the start of the upload,
but the VFS lock is released during the upload, so we need to make again
the check at the end of the upload (with the VFS lock held).