Fix opening notes in a shared directory
If a directory contains a note, and is shared by link, a user with this
link should be able to open the note via the GET /notes/:id/open route.
It was not the case, as the sharecode was not put in the response, but
it is fixed in this commit.