Allow to open a note from the sharing preview (#2433)
When a directory is shared (Cozy to Cozy) with a note inside it, a user
that previews this sharing can open the note. And if the sharing allows
edition, the user will be able to edit the note. The preview sharecode
will only have read permissions, but the stack will generate a
share-interact token so that the user can write on the note.