Try to limit conflicts on sharings
When a folder is shared with replication, and several renaming/move
operations are done inside it, the _revs_diff + _bulk_docs calls on the
directories can be with documents in a "bad" order, which can lead to
conflicts (same name for two directories). We can limit those cases by
delaying the conflict resolution.