Fix cleaning of temporary directories (#2454)
For executing konnectors, services, and mjml, the stack uses nsjail with
a temporary directory. This temporary directory was created inside the
PrepareWorkDir functions, but cleaned via a defer from the function
calling PrepareWorkDir. It means that if a PrepareWorkDir function fails
after creating the temporary directory, neither the PrepareWorkDir
function or its caller will clean this directory. This is fixed by
always creating a function in PrepareWorkDir that is executed by its
caller (this function does nothing before the temporary directory is
created).