Add a skip_redirect_uri_on_authorize parameter (#2463)
The account_types are used when a konnector is using a service with
OAuth 2.0. The OAuth 2.0 danse is started by the stack, with the
parameters from the account_type. Some services have bad implementation
of OAuth 2.0, and a case was reported where the service was not
supporting the redirect_uri parameter. A skip_redirect_uri_on_authorize
parameter was added on the account_types to not send the redirect_uri
for authorize.