Move cookie one level higher on flat subdomains
On flat subdomains, we were using one cookie on the instance domain
(alice.example.com) and one per app (alice-drive.example.com). We had
some issues where those cookies were linked to different active
sessions, which has for consequence that the JWT token that the app got
in its HTML page is linked to one session, and when sent to the stack
for an API request is rejected because it's the other cookie and session
that is used.

To fix those issues, we are now using a single cookie on one level
higher (example.com). As some users are connected on several cozy
instances on the same platform, we have made the cookie name to depend
of the instance to avoid conflicts.

The migration should be transparent, and there are some TODOs to remove
the migration code in a near future (2020-05-15).