Fix index_bad_fullpath on shared directories (#2470)
When a directory is shared from Cozy to Cozy, the stack uses a protocol
inspired by the replication protocol of CouchDB. It means that the new
documents are sent in bulk, and several changes to the same document are
regrouped.

For a tree of directories, the stack has to update the fullpaths of the
descendants when a directory is renamed or moved. This is often done in
the sharing by recomputing the fullpath when the new revision of the
given document is synchronized. But if this happens for a child before
the parent has been renamed/moved on the recipient's instance, it was
failing (index_bad_fullpath was reported by FSCK).

Now, when the parent directory is updated, the stack also recomputes the
fullpaths of the descendants (like the normal indexer of the VFS).