Add a suffix to revoked sharing directories (#2476)
When a sharing is revoked, the directory is kept on the recipients'
instances. If a new sharing is made for the same directory, the
recipients that accept it will have a conflict between the directories
for the old sharing and the new one. It can be confusing. To avoid that,
the directory is renammed when the sharing is revoked (with a suffix).
It will also help people understand that a sharing has been revoked.