Revoke a sharing when its root is trashed
When a sharing is made for a file or folder, and this file or folder is
put to the trash, the sharing should be revoked. Before this commit, it
wasn't the case, and everything inside the shared folder was also put in
the trash, and the synchronisation of the sharing makes the content in
this folder not available for all the members of the sharing. Not good!

Now, the sharing is revoked before the synchronisation can happen, and
the other members of the sharing will keep the content available (ie not
in the trash).