Remove the filepath_operations integration test
This test was useful when it was written. But since then, it has not
helped to catch a regression, and it is unstable on CI (it makes a lot
of document changes in CouchDB, and CouchDB on virtualized environment
does not like this). So, let's remove it to make the CI more stable.