Remove TODOs
These TODOs are here since a long time and nobody has worked on them.
They are not critical and we can safely remove them. Also, I think they
won't be fixed in the coming months, and it's better to track them
somewhere else than in the code. So, let's remove them.