Fix sort for deleted files
From the sharer side, we sort docs before replication in order to apply a bunch of rules, e.g.
directories before files. We also implemented a
[rule](https://github.com/cozy/cozy-stack/commit/4f9a4cdc7134cb586d4461c2b4c05cf3c8a01b8a) on the recipient
side in order to put the deleted files at the end, to efficiently deal
with moves. However, the sort behaviour was not the expected one, and it
was also contradictory with the sort performed by the sharer.
Therefore, we remove this sort on the recipient side and implement it on
the sharer side, with the other rules.