Preserve the redirect parameter on OIDC
When a user tries to connect a mobile app and their instance is on a
context where the login with OIDC is required, it wasn't working because
the OAuth danse was lost during the OIDC authentication. The stack will
now preserve the redirect, which should enable to restart the OAuth
danse after the user has been authenticated.