Add config to allow users to choose their instance with OIDC
When the authentication via OIDC is enabled, the stack needs a way to
check that a user authenticated on the identity provider is really the
owner of the current instance. The stack was able to compute to instance
from the UserInfo response when there is a rule for the instance name at
creation. But, when the user can choose the name of their instance, we
need to rely on another strategy: the stack will take the value of the
oidc_id parameter at the instance creation, and can check it againt the
sub field of the UserInfo response.