Fix updating the io.cozy.shared
When there is a cozy to cozy sharing and the stack updates an
io.cozy.file document with _bulk_docs, it happens that the revisions
chain received is not used with all the revisions. In that case, the
stack mixes the received revisions chain with the local revisions for
_bulk_docs. And we should use the same mixed chain to update the
io.cozy.shared.