Fix revisions tree integrity
When a shared document has updates both from local and remote instances,
the remote revisions might be saved before the local ones in the
revisions tree. If there are several updates, this could cause integrity
issues in the tree. Here is an illustrative scenario:

1. A shared doc has the revision 1-a.
2. The doc is updated twice locally, producing the revisions 2-a and
3-a.
3. The doc is also updated twice on a remote instance, producing the
revisions 4-b and 5-b
4. The revision tree has still the 1-a revision and the following chain is
saved: `[3-a, 4-b, 5-b]`. It corresponds to the last local revision
(3-a) and the incoming remote revisions (4-b, 5-b).
5. The 2-a revision will be later saved in the tree, in a new branch
starting from 1-a.

To solve this, at the step 4, we retrieve the revision history of the
doc, compare the incoming ones and save this revision chain:
`[2-a, 3-a, 4-b, 5-b]`

Avoid long if block