Fix multiple conflicts on the same file (#2559)
When a shared file is updated often by several members of the sharing,
it will lead to conflicts. And if several conflicts happen on the same
file, the stack can try to use a new name for the old content that was
already used. It fails and it aborts the upload of the new content. With
this commit, the stack will try harder to avoid already used names for
conflicts.