Fix OIDC with allow_custom_instance: true
When allow_custom_instance is set to true for a context, the
configuration for userinfo_instance_field is omitted as it is not used.
But the stack was checking that this field was configured and throw an
error when it was not the case on starting the OIDC dance. This commit
relaxes this contraint.