Allow to configure firebase per app (#2601)
* Allow to configure firebase per app

The push notifications can be sent via Firebase to mobile. When we have
several apps developed by several organizations, it can be useful to use
several Firebase accounts. It is now possible. The keys are stored in
the global/io.cozy.account_types database, with a document per app (and
it can be also restricted to a context).

* fixup! Allow to configure firebase per app