Let recipients choose between shortcut or sync
When someone receives a mail for a sharing for the first time, and
accepts the sharing, the stack will ask them if they want to synchronize
the content of the sharing or just add a shortcut to it. For later
sharings, there is no choice as the shortcut has already been added to
their Cozy when they can accept the sharing.