feat: Still continue to to run acount with USER_ACTION_NEEDED.CGU_FORM message
I am not sure of my go syntax but here is what I want :
Accounts with USER_ACTION_NEEDED.CGU_FORM need an action from the user on the website but
it is possible that the user will do this action without knowing this is needed for the connector
and there are some cases the ameli connector where the connector sees the CGU from but not the
user because this form is displayed only when the user connects to his ameli accout, and the
ameli session can be long.

At the moment, I run all the USER_ACTION_NEEDED.CGU_FORM accounts progressively and manually.
I would be nice if the stack would do it automatically and this is the goal of this PR