Change the ownership of the share-by-link permissions
When a permission document is created for a share by link use case, the
stack keeps in the source field the slug of the app that has created the
permission. Later, if the permission is updated or revoked, the stack
was only authorizing the same application to do that. Now, it can also
be another app that has a superset of the given permissions. For
example, if a note is shared from the notes application, it is now
possible to revoke this sharing from Drive.