Fix a panic when refreshing a token
When a recipient is added or removed from a sharing, the stack notifies
the other recipient of this change. With the new possibility to have
shortcut to a sharing, we can have recipients to notify with no OAuth
client. In that case, we need to abort the access token refresh to avoid
a panic.