Fix opening a note in a shared directory (#2691)
When a sharing has been accepted by a member, and then revoked, the
stack may have kept a permission of type share-interact for this
sharing. If the same member accepts the sharing again and creates a note
inside the new shared folder, the other members would have used the
share-interact permission on the old folder, which fails. This commit
fixes this issue by checking that the permission set is correct, and by
updating it if needed.