Make the migrate to layout v3 worker fail fast (#2692)
When there is a copy error on a migration to layout v3 for an instance,
we can abort without waiting to have tried to copy all the other files.

This is done via a combo of:
- a semaphore to limit the number of concurrent calls to Swift
- an errgroup to collect the errors
- a context to coordinate the whole.

When an error occur in a go routine, the errgroup will collect it and
cancel the context. The later calls to aquire the semaphore will then
fail, which will make the VFS walk to abort.