Fix deleting old versions on emptying trash
When the user asks for emptying the trash, the stack puts a job for the
trash worker that will do the real work in background. In particular, it
cleans the old versions. There was a bug that made the bulk requests for
deleting old versions in CouchDB on the wrong database. Let's fix that!
And improve our logs, even if the logs couldn't have given us a clue,
because the CouchDB call on the wrong database is still a success:
deleting documents that don't exist is OK for CouchDB.