Check carbonCopy and electronicSafe metadata
The metadata map for io.cozy.files cannot have a carbonCopy or
electronicSafe key unless the app/konnector that adds it has a
permission on io.cozy.certified.carbonCopy /
io.cozy.certified.electronicSafe.

When a file with such metadata attribute is shared, this attribute will
be removed on the instances of the other members: this attribute is only
valid localy, we don't want to trust other instances as it is easy to
cheat for a self-hosted.