Fix a race condition when using --appdir
When the stack is launched with the --appdir flag, it reads the
permission from the manifest and updates CouchDB when the index page of
an app if served. But, if there are two requests at the same time, it
can lead to a CouchDB conflict, and 500 error for one of those.