Fix CouchDB conflict on instance deletion
When the instance has not been used recently, its indexes/version can be
out of date. The stack updates them automatically when needed, and the
instance deletion can trigger that when checking the accounts to clean.
So, we may have two concurrent updates of the instance document (the
indexes version and the deleting flag), and it can make the deletion
fails with a CouchDB conflict. We have added a retry for this case.