Fix a race condition when importing a Cozy
When a user asks to import a Cozy, the stack (1) creates a job for this
import and (2) puts the importing flag to true on the instance settings.
The import job starts by patching the instance, which means:
- (3) load the previous settings document
- (4) update the instance
- (5) check if the settings document in memory is the same as the one in
  database, and if not try to update it
- (6) other stuff.

The race condition happens when the job starts really fast and (2) is
done after (3). In that case, the step (5) will compare an older
revision of the settings document to a newer one, and will try to put
the older document in CouchDB, which fails with a conflict. This race
condition can be avoided by using `update` instead of `Patch` for the
`Block` and `Unblock` functions, as the stack controls the inputs and
the instance settings document is not involved.