Fix moving a sharing when the two instances are moving
TryTokenForMovedSharing is used when a Cozy has been moved, and a sharing
was not updated on the other Cozy for some reasons. When the other Cozy will
try to make a request to the source Cozy, it will get a 410 Gone error. This
error will also tell it the URL of the new Cozy. Thus, it can try to refresh
the token on the destination Cozy. And, as the refresh token was emitted on
the source Cozy (and not the target Cozy), we need to do some tricks to
manage this refresh. This function is here for that.