Do not import accounts tied to another instance
When moving a Cozy to a new address, the io.cozy.accounts are imported,
except if they are for a konnector that have a hook on the account
deletion. In that case, when the account will be deleted on the source
instance, it will no longer work on the target instance, so it is safer
to not import it.