Fix adding again a "ready" contact to a sharing
When a member of a sharing is in the ready state, and if they are added
again to the sharing, they were added twice to the sharing. But two
members with the same email address confuse the sharing package (the
wrong credentials may be used), so it is better to avoid this
duplication. The owner can still revoke the member from the sharing if
they really want to add it again.