Fix filename for new notes (#2874)
When creating a new note, a call is now made to put the schema. It
updates the updated_at attribute of the file. And this attribute was
used for checking if the file should be renamed when its title has
changed. Now, we are using the created_at to fix notes not renamed
according to their titles.