Do not proxy copy method
The COPY method is not a HTTP standard, and is not supported by Go.
Therefore, we had to translate it to a POST method to support the
related CouchDB API.
To avoid having to translate the POST method to COPY and proxy it to
CouchDB (which is quite ugly), we prefer to add a Copy function in
couchdb.go and direclty query CouchDB from it.
This led to a small refactoring of the makeRequest method to handle
custom headers.