Allow to send the file size via query string
When uploading a file to the VFS, the file size can be sent by clients
via the Content-Length HTTP header. But the desktop client had some
issues with that: Electron removes this header when the body is a stream
or a chunked body, and the client doesn't want to load the whole file in
memory before sending it. So, we have added a fallback where the file
size can also be sent via a Size parameter in the query string.