Send Kdf and kdfIterations on bitwarden login
The bitwarden clients are now expecting the kdf type and iterations on
the login response. This change was made for supporting SSO. The stack
doesn't support that, but we still need to send those fields as the
client logic for deconnection/reconnection is now based on that even
without SSO support.

See https://github.com/bitwarden/jslib/pull/127