[SECU] Fix error handling for getting OAuth client
When requesting the GET /auth/register/:client-id route with no
Authorization header, the stack was sending an error but doesn't stop
processing the request. Thus, the information about the client was also
sent in the response. Ouch!

It means that it was possible to get client_secret from a known
client_id, but this attack alone is not enough to get access to the
personal data of the users: an access_token or refresh_token is still
needed.