Don't see some headers to Swift
When creating a file with the VFS Swift, the stack sends to Swift some
custom meta in the HTTP headers. Swift can respond with a 400 Bad
Request if one of those header is not encoded as it expects (url-encoded
utf-8 NFC, I think but it is not clearly documented). Those headers are
not used currently, so let's just remove them to avoid issues.