Fix running already installed konnectors (afero) (#2924)
When a konnector was installed before we start using brotli for
apps&konnectors assets on a stack with afero FS (not swift), the files
were saved with a .gz extension. They are now saved with a .br
extension.

When running a konnector, we list the files, decompress them and rewrite
them in the temporary directory that will be used to execute them. We
have forgotten to remove the .gz extension for konnectors that were
already installed and nodejs didn't like that as it was expecting a .js
file.