Fix the available apps for intents
When no installed app can be used to answer an intent request, the stack
makes requests to the apps registry to look if other available
applications could do the job. Those requests were overwritting the path
without taking in account the different spaces of apps registy. This
commit fixes this issue by appending to the app registry path.