Make the sharing revocation more robust
When a user puts a shared directory to the trash, the sharing is
revoked. We have seen a case where the revocation failed, which is bad,
and it blocks the user from accepting the sharing again, which is worse.
So, let's make the revocation process more robust, by trying harder to
manage errors and adds a retry loop for removing the shared references.