Fix cleaning app slug in Cozy URL for sharings
When a user wants to accept a sharing, they must give the URL of their
Cozy. There are a few mistakes that can happen easily on this step, and
we want to fix them for the user. For example, the URL can have been
copied-pasted from address bar of the browser, and it contains the app
slug in that case (https://john-drive.mycozy.cloud/ instead of
https://john.mycozy.cloud/). This mistake was already identified and the
stack was fixing the URL for creating the OAuth client, but the fix was
in too low-level code, and the incorrect URL was used later for the
redirection. This commit moves this code to a better place and fix the
issue on redirection.