Fix some issues with sharing + VFS versioning
When files are synchronized via a sharing, the stack use a sharing
indexer to override some behaviors from the classical VFS, like forcing
the revisions to put in CouchDB. And in order to do that, we were using
the UseSharingIndexer method of the VFS that was bugged: the prefix
field was not copied. In most cases, this field wasn't used, but it made
some requests for checking old versions of a file to fail.

In practice, it means that a shared file can have several versions
written in a close time gap, ingoring the mininal delay between versions
from the configuration.