Fix saving docs with OnlyOffice
When an office document was renamed while edited in OnlyOffice, the
callback for saving the file was creating a conflict as the doc revision
was not the expected one due to the renaming. Now, we are also using the
md5sum to detect if we should really create a conflict in such a case.