Allow to redirect to an app for OIDC onboarding
A Cozy can be onboarded with the need to redirect to a specific app.
The password selection page accepts a redirection parameter for this
use case, but when the OIDC method is used for authentication, the user
doesn't go on this page and it doesn't work. We now accept also this
redirection parameter for / when OIDC is enabled.