Fix resharing ciphers with the same member
1. Alice has shared a bitwarden organization + ciphers to Bob
2. Bob has accepted this sharing
3. later, Alice has revoked Bob from the sharing
4. later later, Alice shares again this organization with Bob
5. and Bob accepts

With this scenario, the ciphers in the organization weren't synchronized
on Bob's Cozy. It is due to how the CouchDB replication works. For
example, if a cipher was in revision 3-aaa when Alice has shared it to
Bob the first time, a document was created on Bob's instance with this
revision. Later, when the sharing was revoked, this document was
deleted, and CouchDB will do that by creating a revision 4-bbb. On the
second sharing, the same document is sent with revision 3-aaa. CouchDB
will accept the document with this revision, but the revision 4-bbb is
still seen as a successor of 3-aaa, and as such, the document is still
seen as deleted.

To fix this issue, we haven't tried to update the revision numbers
(really hard to do that in a reliable way), but we are using the same
trick as for files: we are doing a transformation of the ID (XOR), and
the key to do that is changed for each sharing member and when the same
member is accepted again.