Fix moving a cipher from an organization to another
When a cipher is moved from a shared organization to another shared
organization, two share-track jobs are created (one for each
organization). They both tries to update the io.cozy.shared document
for their sharing, but an optimisation for files sharings was preventing
the second job to update the io.cozy.shared. This commit disables this
optimisation for ciphers.