Fix refusing a bitwarden contact
When a user has shared several bitwarden organizations with another
member, the sharings can be in different states for this member:
pending, ready, revoked, etc. For revoking a bitwarden contact, we
should revoke only the sharing with the ready state.

For revoked state, it is already revoked, so no need to do something
more. For pending, we want to let the member the possibility to accept
the sharing and use a new key (so, new fingerprint).