Fix trigger deletion on konnector uninstall
When a konnector is uninstalled, the stack looks at the accounts used by
this konnector, deletes them if the konnector has a on_delete property
in its manifest, and deletes the associated triggers. When there are
several triggers for the same account, only the last one was deleted
because of a bug where the assigned value was lost. This commit fixes
this issue (and I'm surprized that the linter didn't noticed that).