Fix instance deletion with a stale account
When an account still exists but its konnector is missing, the instance
deletion fails (because we can't look at the on_delete field when the
konnector is not installed). We are now trying to reinstall the
konnector in that case, which should help to limit the number of
problematic cases.