Allow to confirm an invited member
When several sharings are made between Alice and Bob, Bob can have
accepted one but not the others. In that case, if Alice confirms
the Bob's identity for the sharing of a bitwarden organization,
cozy-pass-web will also send a confirmation for the other sharings
of bitwarden organizations. For those sharings, Bob can still be
in the invited status, and this commit allow the confirmation to
work in that case.