Fix cleaning OAuth clients with delegated auth (#3187)
When an OAuth client is created and not used in the next hour, it is
deleted in CouchDB by the clean-clients worker. When using delegated
authentication, it was possible to create and use an OAuth client, but
the pending flag was not removed, and the client was removed even if it
had been used. We are removing the flag when giving an access token via
delegated auth.