Allow to use ID token for OIDC authentication (#3181)
The stack can be configured to accept ID token for OIDC authentication,
with the id_token_jwk_url parameter. This parameter will be the URL
where the stack can fetch a JWK which can be used to verify the
signature of the ID token. The ID token must also have a sub field that
matches the oidc_id parameter of the instance for the authentication to
be successful.