Fix opening shared office documents (#3249)
* Fix opening shared office documents

When a sharing member has been invited with several email addresses, and
has accepted and revoked the sharing from one email addresse, and later
accepted the sharing from another email address, their instance will be
listed twice in the sharing (with the status revoked and ready). If this
member creates an office document, and someone else in the sharing tries
to open it, it may fail as the sharing status may have been seen as
revoked (it depends of the order of sharing members). We are now
ignoring the not-ready members to avoid this issue.

* Update model/sharing/open_file.go

Co-authored-by: Sebastien BLAISOT <sebastien.blaisot@cozycloud.cc>

Co-authored-by: Sebastien BLAISOT <sebastien.blaisot@cozycloud.cc>