Fix a race condition on notes
CouchDB is not strongly consistent, and when saving a note with a
changed title, the stack makes two changes: one for renaming the file,
and the other for changing the content. If the user makes a change to
the content, the stack will fetch the document but CouchDB can return
the intermediate version, which can cause trouble (title can be lost).
To avoid this issue, we are pushing the change to the cache, so the
stack will fetch the document from the cache (previously, the cache was
just cleaned).